Remove failure type stuff
FailureType often did not work because HttpClient hides the real failure
from us, so just remove it. Furthermore it was prone to errors.

We now always mark host connections as unchecked when HttpClient
releases a closed connection.

Also, code cleanup in the release method and start plugging more options
here and there.